+++ ops/internal/deployer/state/testdata/v2-intent-output.json
